requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
groq>=0.9.0
chromadb==0.5.23
pypdf==4.3.1
python-docx==1.1.2
watchdog==4.0.2
python-multipart==0.0.12
python-dotenv==1.0.1
sentence-transformers==3.0.1
gdown==5.2.0
httpx==0.27.2
